requirements.txt
anyio
click
fastapi
graphql-core
h11
httptools
idna
psycopg2
pydantic
python-dateutil
python-decouple
python-dotenv
python-multipart
PyYAML
six
sniffio
SQLAlchemy
starlette
strawberry-graphql
typing_extensions
uvicorn
uvloop
watchfiles
websockets


